Updated dismissAnimation
Instead of trying to get a precise transformation of the contentView offscreen, we transform it downward by the height of the UIScreen which should cover all cases (such as when the user swipes the card down off the screen, which would otherwise require a calculation of where the contentView is on the screen during the swipe and a transformation from there. This is little more dirty, but gets the job done on all screen sizes.